+++ projects/igniteui-angular/migrations/update-12_0_0/index.spec.ts
@@ -89,6 +89,51 @@ $theme: igx-avatar-theme(
         );
     });
 
+    it('should replace onValueChange and onValueChanged with valueChange and dragFinished in igx-slider', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/slider.component.html`,
+            `<igx-slider
+            (onValueChange)="someHandler($event)"
+            (onValueChanged)="someHandler($event)"
+            ></igx-slider>`
+        );
+
+        const tree = await schematicRunner.runSchematicAsync(migrationName, {}, appTree).toPromise();
+
+        expect(tree.readContent('/testSrc/appPrefix/component/slider.component.html')).toEqual(
+            `<igx-slider
+            (valueChange)="someHandler($event)"
+            (dragFinished)="someHandler($event)"
+            ></igx-slider>`
+        );
+    });
+
+    it('should replace onProgressChanged with progressChanged in igx-linear-bar', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/linear.component.html`,
+            `<igx-linear-bar [value]="currentValue" (onProgressChanged)="progressChange($event)"></igx-linear-bar>`
+        );
+
+        const tree = await schematicRunner.runSchematicAsync(migrationName, {}, appTree).toPromise();
+
+        expect(tree.readContent('/testSrc/appPrefix/component/linear.component.html')).toEqual(
+            `<igx-linear-bar [value]="currentValue" (progressChanged)="progressChange($event)"></igx-linear-bar>`
+        );
+    });
+
+    it('should replace onProgressChanged with progressChanged in igx-circular-bar', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/circular.component.html`,
+            `<igx-circular-bar [value]="currentValue" (onProgressChanged)="progressChange($event)"></igx-circular-bar>`
+        );
+
+        const tree = await schematicRunner.runSchematicAsync(migrationName, {}, appTree).toPromise();
+
+        expect(tree.readContent('/testSrc/appPrefix/component/circular.component.html')).toEqual(
+            `<igx-circular-bar [value]="currentValue" (progressChanged)="progressChange($event)"></igx-circular-bar>`
+        );
+    });
+
     // IgxTabs
     it('Should update igx-tab-group to igx-tab-item', async () => {
         appTree.create(